pyproject.toml
[tool.poetry]
name = "analystkit"
version = "0.1.0"
description = "Shared Plotly theme, viz helpers, and settings for analytics projects."
authors = ["You <you@example.com>"]
readme = "README.md"
packages = [{ include = "analystkit" }]

[tool.poetry.dependencies]
python = "^3.10"
plotly = "^5.24.0"
pandas = "^2.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
pydantic-settings = "^2.0.0"
# kaleido = "^0.2.1"  # Optional: for static export

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
